Item13619: Make RevCommentPlugin work with Foswiki 2.0
This change untangles the templates a bit from core templates.  It means
that a "revcomment" skin needs to be added to the SKIN setting. Should
be ahead of the "natedit" skin.

Need to follow up with more documentation.  And the default skins
should have the revcomment stuff removed at some point.